# Avoid clone()s & copies in rust-hyper