--- a/rust-hyper/execution-engine/src/op.rs
+++ b/rust-hyper/execution-engine/src/op.rs
@@ -1,11 +1,3 @@
-#[cfg(test)]
-mod tests {
-  #[test]
-  fn can_create_hello_world() {
-    assert_eq!(process([Op::]), estr("hello_world!"));
-  }
-}
-
 use crate::expr;
 
 pub enum Op {