Notes: 
- This will remove source information from errors due to lifetime issues
- `Arc`s replaced by `Rc`s. With more work, It should be possible to reduce `Rc` usage as well and replace with either references or with `Box`s
- implemented serde serializer instead of converting to DOM and then serializing.