@pbiggar 2020-09-27
This is great, performance is improved by about 3x:

Before:

```
  Measuring fizzbuzz
    Reqs/s:       4800.62
    Avg:          20.69ms
    Errors:       N/A
    Slowest:      57.99ms
```

After:

```
rust-hyper/measure_fizzbuzz
    Reqs/s:       12610.21
    Avg:          7.92ms
    Errors:       N/A
    Slowest:      43.13ms
```