@pbiggar 2020-09-27
The benchmark I really want to improve is rust-hyper-async - I currently don't know how to make it compile :(